Fix 64-bit type truncation warnings in GUI coordinate handling
- Change helper_rubber_config and helper_tray_config to use int for x/y/width/height coordinates
- Fix cl_slider.c to explicitly cast long GUI dimensions to int before arithmetic
- Ensure explicit type conversions when passing coordinates to GLUpdateGeometry()
- All GUI coordinates are safely int-sized (physical screen dimensions < INT_MAX)
- Add explicit casts in cl_txtDisp.c for height/width assignments
- Fix type handling in cl_txtLabel.c for xpa_centered() parameters

These changes eliminate implicit 64-to-32 bit truncation in geometry calculations
while maintaining semantic correctness - GUI coordinates are always int in X11.

+++ src/viola/cgen.h
@@ -207,7 +207,7 @@ enum PCodeTypes {
 extern int flag_printPCode;
 extern int flag_printAST;
 
-void printPCode(union PCode* pcode, int* pc, long size);
+void printPCode(union PCode* pcode, long* pc, long size);
 int assignReferences(AST* ast, union PCode* pcode, int* pc);
 int codeGen(AST* ast, union PCode* pcode, int* pc);
